Refactor Allure report generation in Jenkinsfile

+++ Jenkinsfile
@@ -55,15 +55,14 @@ pipeline {
         stage('Generate Allure Report') { 
             steps {
                 script {
-                    sh 'chmod -R 777 $WORKSPACE_DIR/allure-results'
-                    allure([
-                        results: [[path: 'allure-results']]
-                    ])
                     sh '''
-                        chmod -R 777 $WORKSPACE_DIR
-                        rm -f $WORKSPACE_DIR/allure-report.zip
+                        chmod -R 777 $WORKSPACE_DIR/allure-results
+                        allure generate $WORKSPACE_DIR/allure-results --clean
+                        if [ -f $WORKSPACE_DIR/allure-report.zip ]; then
+                            rm -f $WORKSPACE_DIR/allure-report.zip
+                        fi
                         zip -r allure-results-${BUILD_NUMBER}-$(date +"%d-%m-%Y").zip allure-results
-                    '''
+                    '''                
                 }
             }
         }
